refactor(text): consolidate char-boundary helpers into text module
ceil_char_boundary / floor_char_boundary are generic UTF-8 boundary
snapping utilities — stable substitutes for the unstable str methods —
with no injection specificity. CLAUDE.md assigns position/byte
conversions to src/text/, yet they lived in language/injection.rs, and
text/edit.rs carried a third hand-rolled copy of floor_char_boundary.

Move both into src/text/char_boundary.rs as the single home and point
all callers (injection, offset_calculator, range_builder, edit) at it,
removing the duplicate. Pure structural Tidy: identical algorithms,
behavior unchanged.
